requirements.txt
python-dotenv==0.17.1
git+git://github.com/Rapptz/discord.py
dpy-button-utils==1.0.1
expiringdict~=1.2.1
aiohttp~=3.7.4.post0
colorama~=0.4.4
Pillow~=8.2.0